Update Lucene to v3.1.0
This diff updates OMERO from Lucene v2.4.1 to v3.1.0. The most recent
stable version of Lucene (as of this commit) is v4.9.0, however v4.7.2
is that last version to still support Java 6 and v3.6.2 is the latest
version still supported by Hibernate-Search. If/when OMERO drops
support for Java 6, and after Hibernate-Search v5 reaches general
release, we should update Lucene again.

Note by Josh Moore: Using 3.1.0 in order to match Hibernate
                    3.6.10.Final

Aside from updating the library version, the following additional
changes were required for the server to compile:

* Renamed a few imports that were relocated
* Updated `CharTokenizer` subclass to handle `int`s instead of `char`s
* Changed `FullTextAnalyzer` to use updated `createComponents` API
* Added back-compat arguments (`LUCENE_30`) to a few constructors (note:
  this enum name has changed in v4.9.0 to `LUCENE_3_0`)
* Rewrote `SimilarTerms#do` method to use new `FuzzyTermsEnum` technique

Note: With this commit the server and client will build, but will not
run. The reason for this is that the version of Hibernate Search we
build with uses the old strategy for subclassing `Analyzer`.

Conflicts:
	build.xml
	etc/omero.properties

+++ components/server/src/ome/services/fulltext/bridges/ProjectWithImageNameBridge.java
@@ -51,7 +51,7 @@ public class ProjectWithImageNameBridge extends BridgeHelper {
             logger().info("Indexing all image names for " + value);
 
             // Copying lucene options with a new boost value
-            final float reduced_boost = _opts.getBoost().floatValue() / 2;
+            final float reduced_boost = _opts.getBoost() / 2;
             final LuceneOptions opts = new SimpleLuceneOptions(_opts, reduced_boost);
 
             final Project p = (Project) value;
